server/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
deep-translator
